requirements.txt
asttokens==2.4.1
Authlib==1.3.0
beautifulsoup4==4.12.3
blinker==1.7.0
bs4==0.0.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
cryptography==42.0.5
debugpy==1.8.0
decorator==5.1.1
executing==2.0.1
Flask==3.0.1
Flask-Dance==7.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
idna==3.6
ipykernel==6.29.0
ipython==8.20.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.7.1
MarkupSafe==2.1.4
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.6.0
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.9.0.80
packaging==23.2
pandas==2.2.0
parso==0.8.3
pillow==10.2.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
psutil==5.9.8
psycopg2==2.9.9
psycopg2-binary==2.9.9
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
python-dateutil==2.8.2
pytz==2024.1
pywin32==306
pyzmq==25.1.2
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.25
stack-data==0.6.3
sympy==1.12
textblob==0.18.0.post0
threadpoolctl==3.3.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2024.1
Unidecode==1.3.8
urllib3==2.2.0
URLObject==2.4.3
wcwidth==0.2.13
Werkzeug==3.0.1
